Adjusting bootloader combo
Acording to the warnning on this page
https://zmk.dev/docs/features/combos
the bootloader reset is not working as expected on split keyboards when used in combos.

The new behavior uses the combo to apply the adjustment layer as a sticky layer for 2s.
The bootloader key is place on the outer thumb keys in this layer